requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# D:\hewei\soft\check report2\chardeel.py: 5
# D:\hewei\soft\check report2\crnn\CRNN.py: 1
# D:\hewei\soft\check report2\crnn\util.py: 6
# D:\hewei\soft\check report2\main - 区分机种.py: 2,21
# D:\hewei\soft\check report2\main-不区分机种.py: 2
# D:\hewei\soft\check report2\main.py: 2
# D:\hewei\soft\check report2\ocr.py: 2
# D:\hewei\soft\check report2\tifdiv.py: 2
# D:\hewei\soft\check report2\title.py: 9
# D:\hewei\soft\check report2\utils.py: 3
Pillow == 5.4.1

# D:\hewei\soft\check report2\title.py: 1
PyPDF2 == 1.26.0

# D:\hewei\soft\check report2\dbnet\decode.py: 4
Shapely == 1.6.4.post1

# D:\hewei\soft\check report2\analysis_pic.py: 5
# D:\hewei\soft\check report2\title.py: 6
dbnet == 0.2.5

# D:\hewei\soft\check report2\detectTable2.py: 7
matplotlib == 2.1.2

# D:\hewei\soft\check report2\analysis_pic.py: 4
# D:\hewei\soft\check report2\chardeel.py: 4
# D:\hewei\soft\check report2\crnn\CRNN.py: 2
# D:\hewei\soft\check report2\crnn\util.py: 7
# D:\hewei\soft\check report2\dbnet\dbnet_infer.py: 2
# D:\hewei\soft\check report2\dbnet\decode.py: 2
# D:\hewei\soft\check report2\detectTable2.py: 6
# D:\hewei\soft\check report2\title.py: 5
# D:\hewei\soft\check report2\utils.py: 1
numpy == 1.19.0

# D:\hewei\soft\check report2\crnn\CRNN.py: 13
onnxruntime == 1.5.2

# D:\hewei\soft\check report2\analysis_pic.py: 2
# D:\hewei\soft\check report2\chardeel.py: 2
# D:\hewei\soft\check report2\crnn\CRNN.py: 3
# D:\hewei\soft\check report2\dbnet\dbnet_infer.py: 4
# D:\hewei\soft\check report2\dbnet\decode.py: 1
# D:\hewei\soft\check report2\detectTable2.py: 5
# D:\hewei\soft\check report2\main - 区分机种.py: 6
# D:\hewei\soft\check report2\main-不区分机种.py: 6
# D:\hewei\soft\check report2\main.py: 6
# D:\hewei\soft\check report2\title.py: 4
# D:\hewei\soft\check report2\utils.py: 2
opencv_contrib_python == 4.4.0.44

# D:\hewei\soft\check report2\dbnet\decode.py: 3
pyclipper == 1.2.0

# D:\hewei\soft\check report2\title.py: 10
python_Levenshtein == 0.12.0

# D:\hewei\soft\check report2\topdf.py: 3,4,5
reportlab == 3.4.0

# D:\hewei\soft\check report2\dbnet\dbnet_infer.py: 12
utils == 1.0.1
